requirements.txt
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pinecone-client>=2.2.4
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
openai>=1.3.0
transformers>=4.35.0
torch>=2.1.0
PyPDF2>=3.0.1
python-docx>=0.8.11
langchain>=0.0.340
spacy>=3.7.0
pandas>=2.1.0
numpy>=1.24.0
python-multipart>=0.0.6
genai
